Reduce repeated code
Use of sub-personalities is now properly done. No longer uses growing if statements that relied on the personality value being a certain number, rather it is now the name of the personality. At the very least it makes the code more clear.

Also try to prevent trucks from recycling when they are still building something.