Refactor Tactics system
Tactics: The tactics system was quite old and had a few bad practices in it. Now the attacking code is more sane and, more importantly, much more efficient. As such, it may not battle as it used to.
Tactics: The VTOL group now does everything separately in its own timer.
Build: Move buildPhase4() in front of buildDefenses().
Small tweaks elsewhere.