Fix bugs in production and build.
Weapon based units would not be produced in a factory due to bad logic with system units. Now checks that the most basic component of a system can be produced.
Increase power restrictions for building which leads to overall better success.
Slight research path alterations.
and general tweaks in other areas.